requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0  # Required for correlation clustering
xgboost>=2.0.0
lightgbm>=4.0.0

# Data handling
openpyxl>=3.1.0  # For Excel files
pyarrow>=12.0.0  # For efficient data storage

# API & HTTP
aiohttp>=3.9.0  # For async API calls
httpx>=0.25.0
cryptography>=41.0.0  # For RSA signing with API keys

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
black>=23.0.0
flake8>=6.0.0

# Optional: Weather data
# python-nwsapi>=0.1.0  # NWS API wrapper (not available on PyPI - using requests directly)
